code clean up: erased duplicated code
